BotTelegram/requirements.txt
aiohttp               
aiosignal             
APScheduler           
Aspose.Slides         
async                 
async-timeout         
asyncio               
asynctest             
attrs                 
backports.zoneinfo    
certifi               
charset-normalizer    
frozenlist            
idna                  
multidict             
pip                   
pip-install           
pip-plugin-versions   
pyTelegramBotAPI      
pytz                  
pytz-deprecation-shim 
requests              
setuptools            
six                   
telebot               
typing_extensions     
tzdata                
tzlocal               
urllib3               
wheel                 
wincertstore          
yarl                  